requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
scipy
tqdm
Pillow
opencv-python
timm
transformers
mlflow
openslide-python
nystrom-attention
